backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2
python-multipart==0.0.6
google-generativeai==0.3.2
sqlalchemy==2.0.25
aiosqlite==0.19.0
scipy==1.11.4
